requirements.txt
pandas==2.1.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
rich==13.7.0
